requirements.txt
pymupdf
groq
python-dotenv
Pillow
